feat:refactoring code and add test cov
