+++ lib/Model/Sourates/MesSourates.dart
@@ -1,6 +1,3 @@
-// import 'dart:convert';
-// import 'dart:io';
-
 import 'Allsourate.dart';
 
 class MesSourates {
