+++ pubspec.yaml
@@ -16,7 +16,7 @@ dependencies:
   audioplayer_web: ^0.7.1
   share: ^0.6.5+4 #'>=0.6.2+3 <2.0.0'
   arabic_words: ^0.1.22
-  google_fonts: ^1.0.0
+  google_fonts: ^1.0.0 
   intl: ^0.16.1
   flutter_plugin_pdf_viewer: ^1.0.7
   scrollable_positioned_list: ^0.1.5
